[Feat] 스케줄 API 서브

+++ src/main/java/com/example/kotrip/util/algorithm/Node.java
@@ -2,6 +2,8 @@ package com.example.kotrip.util.algorithm;
 
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class Node {
     int curIdx;
@@ -9,10 +11,10 @@ public class Node {
     double latitude;
     double longitude;
 
-    public Node(int curIdx, String name, double latitude, double longitude) {
+    public Node(int curIdx, String name, BigDecimal latitude, BigDecimal longitude) {
         this.curIdx = curIdx;
         this.name = name;
-        this.latitude = latitude;
-        this.longitude = longitude;
+        this.latitude = latitude.doubleValue();
+        this.longitude = longitude.doubleValue();
     }
 }
